Simplify skate.create() to find the constructor and run it. Simpler than document.createElement().
