+++ src/api/create.js
@@ -1,21 +1,6 @@
-import registry from './registry';
+import registry from '../polyfill/registry';
 
-export default function (elementName, parentElementName) {
-  var Ctor = registry.get(parentElementName || elementName);
-  var element;
-
-  if (Ctor) {
-    element = new Ctor();
-  } else {
-    element = document.createElement(elementName);
-
-    // According to the spec, the "is" attribute is always set if a parent
-    // element name is provided even if it has no corresponding registered
-    // custom element.
-    if (parentElementName) {
-      element.setAttribute('is', parentElementName);
-    }
-  }
-
-  return element;
+export default function (name) {
+  var Ctor = registry.get(name);
+  return Ctor && new Ctor() || document.createElement(name);
 }
