+++ src/polyfill/element-constructor.js
@@ -1,17 +1,40 @@
-import createElement from './create-element';
+import {
+  TYPE_ATTRIBUTE,
+  TYPE_CLASS,
+  TYPE_ELEMENT
+} from '../constants';
+
+const DEFAULT_ELEMENT = 'div';
+
+function createElement (options) {
+  var element;
+  var id = options.id;
+  var parent = options.extends;
+  var type = options.type;
+
+  // Allow all types of components to be constructed.
+  if (type === TYPE_ELEMENT) {
+    element = document.createElement(parent || id);
+
+    if (parent) {
+      element.setAttribute('is', id);
+    }
+  } else {
+    element = document.createElement(parent || DEFAULT_ELEMENT);
+
+    if (type === TYPE_ATTRIBUTE) {
+      element.setAttribute(id, '');
+    } else if (type === TYPE_CLASS) {
+      element.className = id;
+    }
+  }
+
+  return element;
+}
 
 export default function (options) {
   function CustomElement () {
-    var element;
-    var elementName = options.id;
-    var parentElementName = options.extends;
-
-    if (parentElementName) {
-      element = createElement(parentElementName);
-      element.setAttribute('is', elementName);
-    } else {
-      element = createElement(elementName);
-    }
+    var element = createElement(options);
 
     // Ensure the definition prototype is up to date with the element's
     // prototype. This ensures that overwriting the element prototype still
