+++ test/unit.js
@@ -1,7 +1,6 @@
 import './boot';
 import './unit/attributes';
-import './unit/components';
-import './unit/create-element';
+import './unit/create';
 import './unit/document-observer';
 import './unit/dom';
 import './unit/events';
