+++ test/unit/create.js
@@ -1,0 +1,10 @@
+'use strict';
+
+import helpers from '../lib/helpers';
+import apiCreate from '../../src/api/create';
+
+describe('skate.create()', function () {
+  function resolved (element) {
+    return element.hasAttribute('resolved');
+  }
+});
